HEXA-523 Quick search issue
Added failing test (skipped for now)
